RPCT-1904, sim: Implementing add for signed integer

+++ LmModel/DeviceEmulator.h
@@ -146,7 +146,7 @@ namespace LmModel
 		//
 	private:
 		quint16 getWord(int wordOffset) const;
-		quint16 getDword(int wordOffset) const;
+		quint32 getDword(int wordOffset) const;
 
 		template <typename TYPE>
 		TYPE getData(int eepromOffset) const;
